Implement test connector bridge
